site-dependent IPR graphs

+++ IPRgeneration/producePedestals.sh
@@ -43,7 +43,6 @@ if [[ $PROD == "PROD5" ]]; then
     TELTYPES=( LST MST-FlashCam MST-NectarCam SST )
 else
     TELTYPES=( LST MST-FlashCam MST-NectarCam SST SCT MAGIC )
-    SITE=( CTA_NORTH CTA_SOUTH CTA_NORTH CTA_SOUTH CTA_SOUTH CTA_NORTH )
 fi
 
 # dedicated scratch directory
@@ -55,41 +54,54 @@ echo "(use this directory as input for all following analysis steps)"
 for i in "${!TELTYPES[@]}"
 do
     T="${TELTYPES[$i]}"
-    echo "Simulating $T for ${PROD}"
 
-    outputFile="${SCRATCH}/pedestals-${T}${MOON}-ze-${ZE}-1k.simtel.gz"
-    rm -f $outputFile
+    for SITE in CTA_NORTH CTA_SOUTH
+    do
+        if [[ $T == "SST" ]] && [[ $SITE == "CTA_NORTH" ]]
+        then
+            continue
+        fi
+        if [[ $T == "MAGIC" ]] && [[ $SITE == "CTA_SOUTH" ]]
+        then
+            continue
+        fi
+        echo "Simulating $T for ${PROD} at $SITE"
 
-    if [[ $PROD == "PROD5" ]] && [[ $T == "SST" ]]; then
-        CFG="${SIM_TELARRAY_PATH}/cfg/CTA/CTA-${PROD}-${T}.cfg"
-    elif [[ $PROD == "PROD5" ]]; then
-        CFG="${SIM_TELARRAY_PATH}/cfg/CTA/CTA-PROD4-${T}.cfg"
-    else
-        CFG="${SIM_TELARRAY_PATH}/cfg/CTA/CTA-${PROD}-${T}.cfg"
-    fi
 
-    INCLUDEOPT=""
-    if [[ $T == "MAGIC" ]]; then
-        CFG="${SIM_TELARRAY_PATH}/cfg/MAGIC/MAGIC1.cfg"
-        INCLUDEOPT="-I${SIM_TELARRAY_PATH}/cfg/MAGIC"
-    fi
+        outputFile="${SCRATCH}/pedestals-${SITE}-${T}${MOON}-ze-${ZE}-1k"
+        rm -f $outputFile
 
-    if [[ $PROD == "PROD5" ]]; then
-        SITEOPT=""
-    else
-        SITEOPT="-D\"${SITE[$i]}\""
-    fi
+        if [[ $PROD == "PROD5" ]] && [[ $T == "SST" ]]; then
+            CFG="${SIM_TELARRAY_PATH}/cfg/CTA/CTA-${PROD}-${T}.cfg"
+        elif [[ $PROD == "PROD5" ]]; then
+            CFG="${SIM_TELARRAY_PATH}/cfg/CTA/CTA-PROD4-${T}.cfg"
+        else
+            CFG="${SIM_TELARRAY_PATH}/cfg/CTA/CTA-${PROD}-${T}.cfg"
+        fi
 
-    ${SIM_TELARRAY_PATH}/bin/sim_telarray -c ${CFG} \
-       -I${SIM_TELARRAY_PATH}/cfg/CTA -I${SIM_TELARRAY_PATH}/cfg/common \
-       -I${SIM_TELARRAY_PATH}/cfg/hess ${INCLUDEOPT} ${SITEOPT} -C Altitude=2150 -C iobuf_maximum=1000000000 \
-       ${MOONOPT} -DNUM_TELESCOPES=1 -C maximum_telescopes=1 \
-       -C atmospheric_transmission=atm_trans_2150_1_10_0_0_2150.dat \
-       -DNSB_AUTOSCALE -C telescope_theta=${ZENITH} -C telescope_phi=180 \
-       -C pedestal_events=1000 \
-       -C output_file=$outputFile \
-       ${CDIR}/dummy1.corsika.gz >& ${SCRATCH}/sim_telarray${MOON}-${T}-ze-${ZE}.log &
+        INCLUDEOPT=""
+        if [[ $T == "MAGIC" ]]; then
+            CFG="${SIM_TELARRAY_PATH}/cfg/MAGIC/MAGIC1.cfg"
+            INCLUDEOPT="-I${SIM_TELARRAY_PATH}/cfg/MAGIC"
+        fi
 
+        if [[ $PROD == "PROD5" ]]; then
+            SITEOPT=""
+        else
+            SITEOPT="-D${SITE}"
+        fi
+
+        ${SIM_TELARRAY_PATH}/bin/sim_telarray -c ${CFG} \
+           -I${SIM_TELARRAY_PATH}/cfg/CTA -I${SIM_TELARRAY_PATH}/cfg/common \
+           -I${SIM_TELARRAY_PATH}/cfg/hess ${INCLUDEOPT} ${SITEOPT} -C Altitude=2150 -C iobuf_maximum=1000000000 \
+           ${MOONOPT} -DNUM_TELESCOPES=1 -C maximum_telescopes=1 \
+           -C atmospheric_transmission=atm_trans_2150_1_10_0_0_2150.dat \
+           -DNSB_AUTOSCALE -C telescope_theta=${ZENITH} -C telescope_phi=180 \
+           -C pedestal_events=1000 \
+           -C output_file=${outputFile}.simtel.gz \
+           ${CDIR}/dummy1.corsika.gz >& ${outputFile}.log
+
+   done
 done
 
 # minor cleanup
